movies/pyproject.toml
[tool.poetry]
name = "movies-admin"
version = "0.1.0"
description = ""
authors = ["aigt"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
Django = "^4.1.1"
django-split-settings = "^1.1.0"
django-debug-toolbar = "^3.6.0"
psycopg2 = "^2.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
